requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
Django==3.1.14
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mongo==0.2.0
packaging==21.3
PyJWT==2.4.0
pymongo==3.12.1
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SQLAlchemy==1.4.36
sqlparse==0.2.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
psycopg2==2.9.1
psycopg2-binary==2.9.1